Delete summary models
Change Clinical Data APIs persistence layer so the interface becomes more verbose
Move chromosome logic from Gene model to GeneServiceImpl
