+++ model/src/main/java/org/cbioportal/model/CancerStudy.java
@@ -1,14 +1,120 @@
 package org.cbioportal.model;
 
-import org.cbioportal.model.summary.CancerStudySummary;
-
 import java.io.Serializable;
 import java.util.Date;
 
-public class CancerStudy extends CancerStudySummary {
+public class CancerStudy implements Serializable {
 
+    private Integer cancerStudyId;
+    private String cancerStudyIdentifier;
+    private String typeOfCancerId;
+    private String name;
+    private String shortName;
+    private String description;
+    private Boolean publicStudy;
+    private String pmid;
+    private String citation;
+    private String groups;
+    private Integer status;
+    private Date importDate;
     private TypeOfCancer typeOfCancer;
     private Integer sampleCount;
+
+    public Integer getCancerStudyId() {
+        return cancerStudyId;
+    }
+
+    public void setCancerStudyId(Integer cancerStudyId) {
+        this.cancerStudyId = cancerStudyId;
+    }
+
+    public String getCancerStudyIdentifier() {
+        return cancerStudyIdentifier;
+    }
+
+    public void setCancerStudyIdentifier(String cancerStudyIdentifier) {
+        this.cancerStudyIdentifier = cancerStudyIdentifier;
+    }
+
+    public String getTypeOfCancerId() {
+        return typeOfCancerId;
+    }
+
+    public void setTypeOfCancerId(String typeOfCancerId) {
+        this.typeOfCancerId = typeOfCancerId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getShortName() {
+        return shortName;
+    }
+
+    public void setShortName(String shortName) {
+        this.shortName = shortName;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Boolean getPublicStudy() {
+        return publicStudy;
+    }
+
+    public void setPublicStudy(Boolean publicStudy) {
+        this.publicStudy = publicStudy;
+    }
+
+    public String getPmid() {
+        return pmid;
+    }
+
+    public void setPmid(String pmid) {
+        this.pmid = pmid;
+    }
+
+    public String getCitation() {
+        return citation;
+    }
+
+    public void setCitation(String citation) {
+        this.citation = citation;
+    }
+
+    public String getGroups() {
+        return groups;
+    }
+
+    public void setGroups(String groups) {
+        this.groups = groups;
+    }
+
+    public Integer getStatus() {
+        return status;
+    }
+
+    public void setStatus(Integer status) {
+        this.status = status;
+    }
+
+    public Date getImportDate() {
+        return importDate;
+    }
+
+    public void setImportDate(Date importDate) {
+        this.importDate = importDate;
+    }
 
     public TypeOfCancer getTypeOfCancer() {
         return typeOfCancer;
