+++ model/src/main/java/org/cbioportal/model/Gene.java
@@ -1,14 +1,57 @@
 package org.cbioportal.model;
 
-import org.cbioportal.model.summary.GeneSummary;
-
+import java.io.Serializable;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-public class Gene extends GeneSummary {
+public class Gene implements Serializable {
 
+    private Integer entrezGeneId;
+    private String hugoGeneSymbol;
+    private String type;
+    private String cytoband;
+    private Integer length;
     private List<String> aliases;
+    private String chromosome;
+
+    public Integer getEntrezGeneId() {
+        return entrezGeneId;
+    }
+
+    public void setEntrezGeneId(Integer entrezGeneId) {
+        this.entrezGeneId = entrezGeneId;
+    }
+
+    public String getHugoGeneSymbol() {
+        return hugoGeneSymbol;
+    }
+
+    public void setHugoGeneSymbol(String hugoGeneSymbol) {
+        this.hugoGeneSymbol = hugoGeneSymbol;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getCytoband() {
+        return cytoband;
+    }
+
+    public void setCytoband(String cytoband) {
+        this.cytoband = cytoband;
+    }
+
+    public Integer getLength() {
+        return length;
+    }
+
+    public void setLength(Integer length) {
+        this.length = length;
+    }
 
     public List<String> getAliases() {
         return aliases;
@@ -19,23 +62,10 @@ public class Gene extends GeneSummary {
     }
 
     public String getChromosome() {
+        return chromosome;
+    }
 
-        if (getCytoband() == null) {
-            return null;
-        }
-        if (getCytoband().toUpperCase().startsWith("X")) {
-            return "X";
-        }
-        if (getCytoband().toUpperCase().startsWith("Y")) {
-            return "Y";
-        }
-
-        Pattern p = Pattern.compile("([0-9]+).*");
-        Matcher m = p.matcher(getCytoband());
-        if (m.find()) {
-            return m.group(1);
-        }
-
-        return null;
+    public void setChromosome(String chromosome) {
+        this.chromosome = chromosome;
     }
 }
