+++ model/src/main/java/org/cbioportal/model/Patient.java
@@ -1,10 +1,46 @@
 package org.cbioportal.model;
 
-import org.cbioportal.model.summary.PatientSummary;
+import java.io.Serializable;
 
-public class Patient extends PatientSummary {
+public class Patient implements Serializable {
 
+    private Integer internalId;
+    private String stableId;
+    private Integer cancerStudyId;
+    private String cancerStudyIdentifier;
     private CancerStudy cancerStudy;
+
+    public Integer getInternalId() {
+        return internalId;
+    }
+
+    public void setInternalId(Integer internalId) {
+        this.internalId = internalId;
+    }
+
+    public String getStableId() {
+        return stableId;
+    }
+
+    public void setStableId(String stableId) {
+        this.stableId = stableId;
+    }
+
+    public Integer getCancerStudyId() {
+        return cancerStudyId;
+    }
+
+    public void setCancerStudyId(Integer cancerStudyId) {
+        this.cancerStudyId = cancerStudyId;
+    }
+
+    public String getCancerStudyIdentifier() {
+        return cancerStudyIdentifier;
+    }
+
+    public void setCancerStudyIdentifier(String cancerStudyIdentifier) {
+        this.cancerStudyIdentifier = cancerStudyIdentifier;
+    }
 
     public CancerStudy getCancerStudy() {
         return cancerStudy;
