+++ model/src/main/java/org/cbioportal/model/Sample.java
@@ -1,8 +1,8 @@
 package org.cbioportal.model;
 
-import org.cbioportal.model.summary.SampleSummary;
+import java.io.Serializable;
 
-public class Sample extends SampleSummary {
+public class Sample implements Serializable {
 
     public enum SampleType {
 
@@ -42,7 +42,61 @@ public class Sample extends SampleSummary {
         }
     }
 
+    private Integer internalId;
+    private String stableId;
+    private SampleType sampleType;
+    private Integer patientId;
+    private String patientStableId;
+    private String typeOfCancerId;
     private Patient patient;
+
+    public Integer getInternalId() {
+        return internalId;
+    }
+
+    public void setInternalId(Integer internalId) {
+        this.internalId = internalId;
+    }
+
+    public String getStableId() {
+        return stableId;
+    }
+
+    public void setStableId(String stableId) {
+        this.stableId = stableId;
+    }
+
+    public SampleType getSampleType() {
+        return sampleType;
+    }
+
+    public void setSampleType(SampleType sampleType) {
+        this.sampleType = sampleType;
+    }
+
+    public Integer getPatientId() {
+        return patientId;
+    }
+
+    public void setPatientId(Integer patientId) {
+        this.patientId = patientId;
+    }
+
+    public String getPatientStableId() {
+        return patientStableId;
+    }
+
+    public void setPatientStableId(String patientStableId) {
+        this.patientStableId = patientStableId;
+    }
+
+    public String getTypeOfCancerId() {
+        return typeOfCancerId;
+    }
+
+    public void setTypeOfCancerId(String typeOfCancerId) {
+        this.typeOfCancerId = typeOfCancerId;
+    }
 
     public Patient getPatient() {
         return patient;
