--- model/src/main/java/org/cbioportal/model/summary/GeneSummary.java
@@ -1,52 +1,0 @@
-package org.cbioportal.model.summary;
-
-import java.io.Serializable;
-
-public abstract class GeneSummary implements Serializable {
-
-    private Integer entrezGeneId;
-    private String hugoGeneSymbol;
-    private String type;
-    private String cytoband;
-    private Integer length;
-
-    public Integer getEntrezGeneId() {
-        return entrezGeneId;
-    }
-
-    public void setEntrezGeneId(Integer entrezGeneId) {
-        this.entrezGeneId = entrezGeneId;
-    }
-
-    public String getHugoGeneSymbol() {
-        return hugoGeneSymbol;
-    }
-
-    public void setHugoGeneSymbol(String hugoGeneSymbol) {
-        this.hugoGeneSymbol = hugoGeneSymbol;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getCytoband() {
-        return cytoband;
-    }
-
-    public void setCytoband(String cytoband) {
-        this.cytoband = cytoband;
-    }
-
-    public Integer getLength() {
-        return length;
-    }
-
-    public void setLength(Integer length) {
-        this.length = length;
-    }
-}
