--- model/src/main/java/org/cbioportal/model/summary/GeneticDataSummary.java
@@ -1,61 +1,0 @@
-package org.cbioportal.model.summary;
-
-import java.io.Serializable;
-
-public abstract class GeneticDataSummary implements Serializable {
-
-    private Integer geneticProfileId;
-    private String geneticProfileStableId;
-    private Integer entrezGeneId;
-    private Integer sampleId;
-    private String sampleStableId;
-    private String value;
-
-    public Integer getGeneticProfileId() {
-        return geneticProfileId;
-    }
-
-    public void setGeneticProfileId(Integer geneticProfileId) {
-        this.geneticProfileId = geneticProfileId;
-    }
-
-    public String getGeneticProfileStableId() {
-        return geneticProfileStableId;
-    }
-
-    public void setGeneticProfileStableId(String geneticProfileStableId) {
-        this.geneticProfileStableId = geneticProfileStableId;
-    }
-
-    public Integer getEntrezGeneId() {
-        return entrezGeneId;
-    }
-
-    public void setEntrezGeneId(Integer entrezGeneId) {
-        this.entrezGeneId = entrezGeneId;
-    }
-
-    public Integer getSampleId() {
-        return sampleId;
-    }
-
-    public void setSampleId(Integer sampleId) {
-        this.sampleId = sampleId;
-    }
-
-    public String getSampleStableId() {
-        return sampleStableId;
-    }
-
-    public void setSampleStableId(String sampleStableId) {
-        this.sampleStableId = sampleStableId;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-}
