--- model/src/main/java/org/cbioportal/model/summary/GeneticProfileSummary.java
@@ -1,88 +1,0 @@
-package org.cbioportal.model.summary;
-
-import java.io.Serializable;
-
-public abstract class GeneticProfileSummary implements Serializable {
-
-    private Integer geneticProfileId;
-    private String stableId;
-    private Integer cancerStudyId;
-    private String cancerStudyIdentifier;
-    private String geneticAlterationType;
-    private String datatype;
-    private String name;
-    private String description;
-    private Boolean showProfileInAnalysisTab;
-
-    public Integer getGeneticProfileId() {
-        return geneticProfileId;
-    }
-
-    public void setGeneticProfileId(Integer geneticProfileId) {
-        this.geneticProfileId = geneticProfileId;
-    }
-
-    public String getStableId() {
-        return stableId;
-    }
-
-    public void setStableId(String stableId) {
-        this.stableId = stableId;
-    }
-
-    public Integer getCancerStudyId() {
-        return cancerStudyId;
-    }
-
-    public void setCancerStudyId(Integer cancerStudyId) {
-        this.cancerStudyId = cancerStudyId;
-    }
-
-    public String getCancerStudyIdentifier() {
-        return cancerStudyIdentifier;
-    }
-
-    public void setCancerStudyIdentifier(String cancerStudyIdentifier) {
-        this.cancerStudyIdentifier = cancerStudyIdentifier;
-    }
-
-    public String getGeneticAlterationType() {
-        return geneticAlterationType;
-    }
-
-    public void setGeneticAlterationType(String geneticAlterationType) {
-        this.geneticAlterationType = geneticAlterationType;
-    }
-
-    public String getDatatype() {
-        return datatype;
-    }
-
-    public void setDatatype(String datatype) {
-        this.datatype = datatype;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Boolean getShowProfileInAnalysisTab() {
-        return showProfileInAnalysisTab;
-    }
-
-    public void setShowProfileInAnalysisTab(Boolean showProfileInAnalysisTab) {
-        this.showProfileInAnalysisTab = showProfileInAnalysisTab;
-    }
-}
