--- model/src/main/java/org/cbioportal/model/summary/MutationSummary.java
@@ -1,277 +1,0 @@
-package org.cbioportal.model.summary;
-
-import java.io.Serializable;
-
-public abstract class MutationSummary implements Serializable {
-
-    private Integer mutationEventId;
-    private Integer geneticProfileId;
-    private String geneticProfileStableId;
-    private Integer sampleId;
-    private String sampleStableId;
-    private Integer entrezGeneId;
-    private String center;
-    private String sequencer;
-    private String mutationStatus;
-    private String validationStatus;
-    private String tumorSeqAllele1;
-    private String tumorSeqAllele2;
-    private String matchedNormSampleBarcode;
-    private String matchNormSeqAllele1;
-    private String matchNormSeqAllele2;
-    private String tumorValidationAllele1;
-    private String tumorValidationAllele2;
-    private String matchNormValidationAllele1;
-    private String matchNormValidationAllele2;
-    private String verificationStatus;
-    private String sequencingPhase;
-    private String sequenceSource;
-    private String validationMethod;
-    private String score;
-    private String bamFile;
-    private Integer tumorAltCount;
-    private Integer tumorRefCount;
-    private Integer normalAltCount;
-    private Integer normalRefCount;
-    private String aminoAcidChange;
-
-    public Integer getMutationEventId() {
-        return mutationEventId;
-    }
-
-    public void setMutationEventId(Integer mutationEventId) {
-        this.mutationEventId = mutationEventId;
-    }
-
-    public Integer getGeneticProfileId() {
-        return geneticProfileId;
-    }
-
-    public void setGeneticProfileId(Integer geneticProfileId) {
-        this.geneticProfileId = geneticProfileId;
-    }
-
-    public String getGeneticProfileStableId() {
-        return geneticProfileStableId;
-    }
-
-    public void setGeneticProfileStableId(String geneticProfileStableId) {
-        this.geneticProfileStableId = geneticProfileStableId;
-    }
-
-    public Integer getSampleId() {
-        return sampleId;
-    }
-
-    public void setSampleId(Integer sampleId) {
-        this.sampleId = sampleId;
-    }
-
-    public String getSampleStableId() {
-        return sampleStableId;
-    }
-
-    public void setSampleStableId(String sampleStableId) {
-        this.sampleStableId = sampleStableId;
-    }
-
-    public Integer getEntrezGeneId() {
-        return entrezGeneId;
-    }
-
-    public void setEntrezGeneId(Integer entrezGeneId) {
-        this.entrezGeneId = entrezGeneId;
-    }
-
-    public String getCenter() {
-        return center;
-    }
-
-    public void setCenter(String center) {
-        this.center = center;
-    }
-
-    public String getSequencer() {
-        return sequencer;
-    }
-
-    public void setSequencer(String sequencer) {
-        this.sequencer = sequencer;
-    }
-
-    public String getMutationStatus() {
-        return mutationStatus;
-    }
-
-    public void setMutationStatus(String mutationStatus) {
-        this.mutationStatus = mutationStatus;
-    }
-
-    public String getValidationStatus() {
-        return validationStatus;
-    }
-
-    public void setValidationStatus(String validationStatus) {
-        this.validationStatus = validationStatus;
-    }
-
-    public String getTumorSeqAllele1() {
-        return tumorSeqAllele1;
-    }
-
-    public void setTumorSeqAllele1(String tumorSeqAllele1) {
-        this.tumorSeqAllele1 = tumorSeqAllele1;
-    }
-
-    public String getTumorSeqAllele2() {
-        return tumorSeqAllele2;
-    }
-
-    public void setTumorSeqAllele2(String tumorSeqAllele2) {
-        this.tumorSeqAllele2 = tumorSeqAllele2;
-    }
-
-    public String getMatchedNormSampleBarcode() {
-        return matchedNormSampleBarcode;
-    }
-
-    public void setMatchedNormSampleBarcode(String matchedNormSampleBarcode) {
-        this.matchedNormSampleBarcode = matchedNormSampleBarcode;
-    }
-
-    public String getMatchNormSeqAllele1() {
-        return matchNormSeqAllele1;
-    }
-
-    public void setMatchNormSeqAllele1(String matchNormSeqAllele1) {
-        this.matchNormSeqAllele1 = matchNormSeqAllele1;
-    }
-
-    public String getMatchNormSeqAllele2() {
-        return matchNormSeqAllele2;
-    }
-
-    public void setMatchNormSeqAllele2(String matchNormSeqAllele2) {
-        this.matchNormSeqAllele2 = matchNormSeqAllele2;
-    }
-
-    public String getTumorValidationAllele1() {
-        return tumorValidationAllele1;
-    }
-
-    public void setTumorValidationAllele1(String tumorValidationAllele1) {
-        this.tumorValidationAllele1 = tumorValidationAllele1;
-    }
-
-    public String getTumorValidationAllele2() {
-        return tumorValidationAllele2;
-    }
-
-    public void setTumorValidationAllele2(String tumorValidationAllele2) {
-        this.tumorValidationAllele2 = tumorValidationAllele2;
-    }
-
-    public String getMatchNormValidationAllele1() {
-        return matchNormValidationAllele1;
-    }
-
-    public void setMatchNormValidationAllele1(String matchNormValidationAllele1) {
-        this.matchNormValidationAllele1 = matchNormValidationAllele1;
-    }
-
-    public String getMatchNormValidationAllele2() {
-        return matchNormValidationAllele2;
-    }
-
-    public void setMatchNormValidationAllele2(String matchNormValidationAllele2) {
-        this.matchNormValidationAllele2 = matchNormValidationAllele2;
-    }
-
-    public String getVerificationStatus() {
-        return verificationStatus;
-    }
-
-    public void setVerificationStatus(String verificationStatus) {
-        this.verificationStatus = verificationStatus;
-    }
-
-    public String getSequencingPhase() {
-        return sequencingPhase;
-    }
-
-    public void setSequencingPhase(String sequencingPhase) {
-        this.sequencingPhase = sequencingPhase;
-    }
-
-    public String getSequenceSource() {
-        return sequenceSource;
-    }
-
-    public void setSequenceSource(String sequenceSource) {
-        this.sequenceSource = sequenceSource;
-    }
-
-    public String getValidationMethod() {
-        return validationMethod;
-    }
-
-    public void setValidationMethod(String validationMethod) {
-        this.validationMethod = validationMethod;
-    }
-
-    public String getScore() {
-        return score;
-    }
-
-    public void setScore(String score) {
-        this.score = score;
-    }
-
-    public String getBamFile() {
-        return bamFile;
-    }
-
-    public void setBamFile(String bamFile) {
-        this.bamFile = bamFile;
-    }
-
-    public Integer getTumorAltCount() {
-        return tumorAltCount;
-    }
-
-    public void setTumorAltCount(Integer tumorAltCount) {
-        this.tumorAltCount = tumorAltCount;
-    }
-
-    public Integer getTumorRefCount() {
-        return tumorRefCount;
-    }
-
-    public void setTumorRefCount(Integer tumorRefCount) {
-        this.tumorRefCount = tumorRefCount;
-    }
-
-    public Integer getNormalAltCount() {
-        return normalAltCount;
-    }
-
-    public void setNormalAltCount(Integer normalAltCount) {
-        this.normalAltCount = normalAltCount;
-    }
-
-    public Integer getNormalRefCount() {
-        return normalRefCount;
-    }
-
-    public void setNormalRefCount(Integer normalRefCount) {
-        this.normalRefCount = normalRefCount;
-    }
-
-    public String getAminoAcidChange() {
-        return aminoAcidChange;
-    }
-
-    public void setAminoAcidChange(String aminoAcidChange) {
-        this.aminoAcidChange = aminoAcidChange;
-    }
-}
