--- model/src/main/java/org/cbioportal/model/summary/PatientSummary.java
@@ -1,43 +1,0 @@
-package org.cbioportal.model.summary;
-
-import java.io.Serializable;
-
-public abstract class PatientSummary implements Serializable {
-
-    private Integer internalId;
-    private String stableId;
-    private Integer cancerStudyId;
-    private String cancerStudyIdentifier;
-
-    public Integer getInternalId() {
-        return internalId;
-    }
-
-    public void setInternalId(Integer internalId) {
-        this.internalId = internalId;
-    }
-
-    public String getStableId() {
-        return stableId;
-    }
-
-    public void setStableId(String stableId) {
-        this.stableId = stableId;
-    }
-
-    public Integer getCancerStudyId() {
-        return cancerStudyId;
-    }
-
-    public void setCancerStudyId(Integer cancerStudyId) {
-        this.cancerStudyId = cancerStudyId;
-    }
-
-    public String getCancerStudyIdentifier() {
-        return cancerStudyIdentifier;
-    }
-
-    public void setCancerStudyIdentifier(String cancerStudyIdentifier) {
-        this.cancerStudyIdentifier = cancerStudyIdentifier;
-    }
-}
