+++ persistence/persistence-api/src/main/java/org/cbioportal/persistence/PersistenceConstants.java
@@ -3,4 +3,5 @@ package org.cbioportal.persistence;
 public class PersistenceConstants {
 
     public static final String DETAILED_PROJECTION = "DETAILED";
+    public static final String SAMPLE_CLINICAL_DATA_TYPE = "SAMPLE";
 }
