+++ persistence/persistence-mybatis-test/src/test/java/org/cbioportal/persistence/mybatis/ClinicalDataMyBatisRepositoryTest.java
@@ -1,7 +1,8 @@
 package org.cbioportal.persistence.mybatis;
 
-import org.cbioportal.model.*;
+import org.cbioportal.model.ClinicalData;
 import org.cbioportal.model.meta.BaseMeta;
+import org.cbioportal.persistence.PersistenceConstants;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -24,178 +25,59 @@ public class ClinicalDataMyBatisRepositoryTest {
     @Test
     public void getAllClinicalDataOfSampleInStudyEmptyResult() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("invalid_study");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, null, null, "ID", null, null,
-                        null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy("invalid_study",
+                null, null, "ID", null, null, null, null);
 
         Assert.assertEquals(0, result.size());
     }
 
     @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudyNullSampleNullAttributeIdProjection() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, null, null, "ID", null, null,
-                        null, null);
-
-        Assert.assertEquals(6, result.size());
-        SampleClinicalData data = result.get(0);
-        Assert.assertEquals("DAYS_TO_COLLECTION", data.getAttrId());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getSample());
-        Assert.assertNull(data.getAttrValue());
-    }
-
-    @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudyNullSampleNullAttributeSummaryProjection()
+    public void getAllClinicalDataOfSampleInStudyNullAttributeSummaryProjection()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, null, null, "SUMMARY", null,
-                        null, null, null);
-
-        Assert.assertEquals(6, result.size());
-        SampleClinicalData data = result.get(0);
-        Assert.assertEquals("DAYS_TO_COLLECTION", data.getAttrId());
-        Assert.assertEquals("276", data.getAttrValue());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getSample());
-    }
-
-    @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudyNullSampleNullAttributeDetailedProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, null, null, "DETAILED", null,
-                        null, null, null);
-
-        Assert.assertEquals(6, result.size());
-        SampleClinicalData data = result.get(0);
-        Assert.assertEquals("OTHER_SAMPLE_ID", data.getAttrId());
-        Assert.assertEquals("5C631CE8-F96A-4C35-A459-556FC4AB21E1", data.getAttrValue());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        ClinicalAttribute clinicalAttribute = data.getClinicalAttribute();
-        Assert.assertEquals("OTHER_SAMPLE_ID", clinicalAttribute.getAttrId());
-        Assert.assertEquals("Legacy DMP sample identifier (DMPnnnn)", clinicalAttribute.getDescription());
-        Assert.assertEquals("STRING", clinicalAttribute.getDatatype());
-        Assert.assertEquals("Other Sample ID", clinicalAttribute.getDisplayName());
-        Assert.assertEquals(false, clinicalAttribute.getPatientAttribute());
-        Assert.assertEquals("1", clinicalAttribute.getPriority());
-    }
-
-    @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudySingleSampleNullAttributeSummaryProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        null, null, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB-01", null, "SUMMARY", null, null, null, null);
 
         Assert.assertEquals(3, result.size());
-        SampleClinicalData data = result.get(0);
+        ClinicalData data = result.get(0);
         Assert.assertEquals("DAYS_TO_COLLECTION", data.getAttrId());
         Assert.assertEquals("276", data.getAttrValue());
         Assert.assertEquals((Integer) 1, data.getInternalId());
         Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getSample());
     }
 
     @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudySingleSampleWithAttributeSummaryProjection()
+    public void getAllClinicalDataOfSampleInStudyWithAttributeSummaryProjection()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, "OTHER_SAMPLE_ID",
-                        "SUMMARY", null, null, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB-01", "OTHER_SAMPLE_ID", "SUMMARY", null, null, null, null);
 
         Assert.assertEquals(1, result.size());
-        SampleClinicalData data = result.get(0);
+        ClinicalData data = result.get(0);
         Assert.assertEquals("OTHER_SAMPLE_ID", data.getAttrId());
         Assert.assertEquals("5C631CE8-F96A-4C35-A459-556FC4AB21E1", data.getAttrValue());
         Assert.assertEquals((Integer) 1, data.getInternalId());
         Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getSample());
     }
 
     @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudyMultipleSampleNullAttributeSummaryProjection()
+    public void getAllClinicalDataOfSampleInStudyNullAttributeSummaryProjection2PageSize()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        sampleIds.add("TCGA-A1-A0SD-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        null, null, null, null);
-
-        Assert.assertEquals(6, result.size());
-    }
-
-    @Test
-    public void getAllClinicalDataOfSampleInStudyMultipleStudyMultipleSampleNullAttributeSummaryProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        studyIds.add("acc_tcga");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        sampleIds.add("TCGA-A1-B0SO-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        null, null, null, null);
-
-        Assert.assertEquals(5, result.size());
-    }
-
-    @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudySingleSampleNullAttributeSummaryProjection2PageSize()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        2, 0, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB-01", null, "SUMMARY", 2, 0, null, null);
 
         Assert.assertEquals(2, result.size());
     }
 
     @Test
-    public void getAllClinicalDataOfSampleInStudySingleStudySingleSampleNullAttributeSummaryProjectionAttrIdSort()
+    public void getAllClinicalDataOfSampleInStudyNullAttributeSummaryProjectionAttrIdSort()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        List<SampleClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        null, null, "attrId", "ASC");
+        List<ClinicalData> result =
+                clinicalDataMyBatisRepository.getAllClinicalDataOfSampleInStudy("study_tcga_pub", "TCGA-A1-A0SB-01",
+                        null, "SUMMARY", null, null, "attrId", "ASC");
 
         Assert.assertEquals(3, result.size());
         Assert.assertEquals("DAYS_TO_COLLECTION", result.get(0).getAttrId());
@@ -206,235 +88,83 @@ public class ClinicalDataMyBatisRepositoryTest {
     @Test
     public void getMetaSampleClinicalDataZeroCount() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("invalid_study");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData(studyIds, null, null);
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData("invalid_study", null, null);
         Assert.assertEquals((Integer) 0, result.getTotalCount());
     }
 
     @Test
-    public void getMetaSampleClinicalDataSingleStudyNullSampleNullAttribute() throws Exception {
+    public void getMetaSampleClinicalDataNullAttribute() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData(studyIds, null, null);
-        Assert.assertEquals((Integer) 6, result.getTotalCount());
-    }
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData("study_tcga_pub", "TCGA-A1-A0SB-01",
+                null);
 
-    @Test
-    public void getMetaSampleClinicalDataSingleStudySingleSampleNullAttribute() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData(studyIds, sampleIds, null);
         Assert.assertEquals((Integer) 3, result.getTotalCount());
     }
 
     @Test
-    public void getMetaSampleClinicalDataSingleStudySingleSampleWithAttribute() throws Exception {
+    public void getMetaSampleClinicalDataWithAttribute() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData(studyIds, sampleIds,
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData("study_tcga_pub", "TCGA-A1-A0SB-01",
                 "OTHER_SAMPLE_ID");
 
         Assert.assertEquals((Integer) 1, result.getTotalCount());
     }
 
     @Test
-    public void getMetaSampleClinicalDataMultipleStudyMultipleSampleNullAttribute() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        studyIds.add("acc_tcga");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB-01");
-        sampleIds.add("TCGA-A1-B0SO-01");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaSampleClinicalData(studyIds, sampleIds, null);
-
-        Assert.assertEquals((Integer) 5, result.getTotalCount());
-    }
-
-    @Test
     public void getAllClinicalDataOfPatientInStudyEmptyResult() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("invalid_study");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, null, null, "ID", null, null,
-                        null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy("invalid_study",
+                null, null, "ID", null, null, null, null);
 
         Assert.assertEquals(0, result.size());
     }
 
     @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudyNullPatientNullAttributeIdProjection() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, null, null, "ID", null, null,
-                        null, null);
-
-        Assert.assertEquals(4, result.size());
-        PatientClinicalData data = result.get(0);
-        Assert.assertEquals("FORM_COMPLETION_DATE", data.getAttrId());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getPatient());
-        Assert.assertNull(data.getAttrValue());
-    }
-
-    @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudyNullPatientNullAttributeSummaryProjection()
+    public void getAllClinicalDataOfPatientInStudyNullAttributeSummaryProjection()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, null, null, "SUMMARY", null,
-                        null, null, null);
-
-        Assert.assertEquals(4, result.size());
-        PatientClinicalData data = result.get(0);
-        Assert.assertEquals("FORM_COMPLETION_DATE", data.getAttrId());
-        Assert.assertEquals("2013-12-5", data.getAttrValue());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getPatient());
-    }
-
-    @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudyNullPatientNullAttributeDetailedProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, null, null, "DETAILED", null,
-                        null, null, null);
-
-        Assert.assertEquals(4, result.size());
-        PatientClinicalData data = result.get(0);
-        Assert.assertEquals("RETROSPECTIVE_COLLECTION", data.getAttrId());
-        Assert.assertEquals("NO", data.getAttrValue());
-        Assert.assertEquals((Integer) 1, data.getInternalId());
-        ClinicalAttribute clinicalAttribute = data.getClinicalAttribute();
-        Assert.assertEquals("RETROSPECTIVE_COLLECTION", clinicalAttribute.getAttrId());
-        Assert.assertEquals("Text indicator for the time frame of tissue procurement, indicating that the tissue was " +
-                "obtained and stored prior to the initiation of the project.", clinicalAttribute.getDescription());
-        Assert.assertEquals("STRING", clinicalAttribute.getDatatype());
-        Assert.assertEquals("Tissue Retrospective Collection Indicator", clinicalAttribute.getDisplayName());
-        Assert.assertEquals(true, clinicalAttribute.getPatientAttribute());
-        Assert.assertEquals("1", clinicalAttribute.getPriority());
-    }
-
-    @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudySinglePatientNullAttributeSummaryProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds, null, "SUMMARY",
-                        null, null, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB", null, "SUMMARY", null, null, null, null);
 
         Assert.assertEquals(3, result.size());
-        PatientClinicalData data = result.get(0);
+        ClinicalData data = result.get(0);
         Assert.assertEquals("FORM_COMPLETION_DATE", data.getAttrId());
         Assert.assertEquals("2013-12-5", data.getAttrValue());
         Assert.assertEquals((Integer) 1, data.getInternalId());
         Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getPatient());
     }
 
     @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudySinglePatientWithAttributeSummaryProjection()
+    public void getAllClinicalDataOfPatientInStudyWithAttributeSummaryProjection()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds,
-                        "OTHER_PATIENT_ID", "SUMMARY", null, null, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB", "OTHER_PATIENT_ID", "SUMMARY", null, null, null, null);
 
         Assert.assertEquals(1, result.size());
-        PatientClinicalData data = result.get(0);
+        ClinicalData data = result.get(0);
         Assert.assertEquals("OTHER_PATIENT_ID", data.getAttrId());
         Assert.assertEquals("286CF147-B7F7-4A05-8E41-7FBD3717AD71", data.getAttrValue());
         Assert.assertEquals((Integer) 1, data.getInternalId());
         Assert.assertNull(data.getClinicalAttribute());
-        Assert.assertNull(data.getPatient());
     }
 
     @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudyMultiplePatientNullAttributeSummaryProjection()
+    public void getAllClinicalDataOfPatientInStudyNullAttributeSummaryProjection2PageSize()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        patientIds.add("TCGA-A1-A0SD");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds, null, "SUMMARY",
-                        null, null, null, null);
-
-        Assert.assertEquals(4, result.size());
-    }
-
-    @Test
-    public void getAllClinicalDataOfPatientInStudyMultipleStudyMultiplePatientNullAttributeSummaryProjection()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        studyIds.add("acc_tcga");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        patientIds.add("TCGA-A1-B0SO");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds, null, "SUMMARY",
-                        null, null, null, null);
-
-        Assert.assertEquals(7, result.size());
-    }
-
-    @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudySinglePatientNullAttributeSummaryProjection2PageSize()
-            throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds, null, "SUMMARY",
-                        2, 0, null, null);
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB", null, "SUMMARY", 2, 0, null, null);
 
         Assert.assertEquals(2, result.size());
     }
 
     @Test
-    public void getAllClinicalDataOfPatientInStudySingleStudySinglePatientNullAttributeSummaryProjectionAttrIdSort()
+    public void getAllClinicalDataOfPatientInStudyNullAttributeSummaryProjectionAttrIdSort()
             throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add("TCGA-A1-A0SB");
-        List<PatientClinicalData> result =
-                clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy(studyIds, sampleIds, null, "SUMMARY",
-                        null, null, "attrId", "ASC");
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataOfPatientInStudy("study_tcga_pub",
+                "TCGA-A1-A0SB", null, "SUMMARY", null, null, "attrId", "ASC");
 
         Assert.assertEquals(3, result.size());
         Assert.assertEquals("FORM_COMPLETION_DATE", result.get(0).getAttrId());
@@ -445,56 +175,149 @@ public class ClinicalDataMyBatisRepositoryTest {
     @Test
     public void getMetaPatientClinicalDataZeroCount() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("invalid_study");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData(studyIds, null, null);
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData("invalid_study", null, null);
         Assert.assertEquals((Integer) 0, result.getTotalCount());
     }
 
     @Test
-    public void getMetaPatientClinicalDataSingleStudyNullPatientNullAttribute() throws Exception {
+    public void getMetaPatientClinicalDataNullAttribute() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData(studyIds, null, null);
-        Assert.assertEquals((Integer) 4, result.getTotalCount());
-    }
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData("study_tcga_pub", "TCGA-A1-A0SB",
+                null);
 
-    @Test
-    public void getMetaPatientClinicalDataSingleStudySinglePatientNullAttribute() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData(studyIds, patientIds, null);
         Assert.assertEquals((Integer) 3, result.getTotalCount());
     }
 
     @Test
-    public void getMetaPatientClinicalDataSingleStudySinglePatientWithAttribute() throws Exception {
+    public void getMetaPatientClinicalDataWithAttribute() throws Exception {
 
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add("study_tcga_pub");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData(studyIds, patientIds,
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData("study_tcga_pub", "TCGA-A1-A0SB",
                 "OTHER_PATIENT_ID");
 
         Assert.assertEquals((Integer) 1, result.getTotalCount());
     }
 
     @Test
-    public void getMetaPatientClinicalDataMultipleStudyMultiplePatientNullAttribute() throws Exception {
+    public void getAllClinicalDataInStudyEmptyResult() throws Exception {
+
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataInStudy("invalid_study",
+                null, PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "ID", null, null, null, null);
+
+        Assert.assertEquals(0, result.size());
+    }
+
+    @Test
+    public void getAllClinicalDataInStudyNullAttributeSummaryProjection() throws Exception {
+
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataInStudy("study_tcga_pub",
+                null, PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "SUMMARY", null, null, null, null);
+
+        Assert.assertEquals(6, result.size());
+        ClinicalData data = result.get(0);
+        Assert.assertEquals("DAYS_TO_COLLECTION", data.getAttrId());
+        Assert.assertEquals("276", data.getAttrValue());
+        Assert.assertEquals((Integer) 1, data.getInternalId());
+        Assert.assertNull(data.getClinicalAttribute());
+    }
+
+    @Test
+    public void getAllClinicalDataInStudyWithAttributeSummaryProjection() throws Exception {
+
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataInStudy("study_tcga_pub",
+                "DAYS_TO_COLLECTION", PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "SUMMARY", null, null, null,
+                null);
+
+        Assert.assertEquals(1, result.size());
+        ClinicalData data = result.get(0);
+        Assert.assertEquals("DAYS_TO_COLLECTION", data.getAttrId());
+        Assert.assertEquals("276", data.getAttrValue());
+        Assert.assertEquals((Integer) 1, data.getInternalId());
+        Assert.assertNull(data.getClinicalAttribute());
+    }
+
+    @Test
+    public void getAllClinicalDataInStudyNullAttributeSummaryProjection2PageSize() throws Exception {
+
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataInStudy("study_tcga_pub",
+                null, PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "SUMMARY", 2, 0, null, null);
+
+        Assert.assertEquals(2, result.size());
+    }
+
+    @Test
+    public void getAllClinicalDataInStudyNullAttributeSummaryProjectionAttrIdSort() throws Exception {
+
+        List<ClinicalData> result = clinicalDataMyBatisRepository.getAllClinicalDataInStudy("study_tcga_pub",
+                null, PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "SUMMARY", null, null, "attrId", "ASC");
+
+        Assert.assertEquals(6, result.size());
+        Assert.assertEquals("DAYS_TO_COLLECTION", result.get(0).getAttrId());
+        Assert.assertEquals("IS_FFPE", result.get(1).getAttrId());
+        Assert.assertEquals("OCT_EMBEDDED", result.get(2).getAttrId());
+        Assert.assertEquals("OTHER_SAMPLE_ID", result.get(3).getAttrId());
+        Assert.assertEquals("PATHOLOGY_REPORT_FILE_NAME", result.get(4).getAttrId());
+        Assert.assertEquals("SAMPLE_TYPE", result.get(5).getAttrId());
+    }
+
+    @Test
+    public void getMetaAllClinicalDataZeroCount() throws Exception {
+
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaAllClinicalData("invalid_study", null,
+                PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE);
+
+        Assert.assertEquals((Integer) 0, result.getTotalCount());
+    }
+
+    @Test
+    public void getMetaAllClinicalDataNullAttribute() throws Exception {
+
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaAllClinicalData("study_tcga_pub", null,
+                PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE);
+
+        Assert.assertEquals((Integer) 6, result.getTotalCount());
+    }
+
+    @Test
+    public void getMetaAllClinicalDataWithAttribute() throws Exception {
+
+        BaseMeta result = clinicalDataMyBatisRepository.getMetaAllClinicalData("study_tcga_pub", "DAYS_TO_COLLECTION",
+                PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE);
+
+        Assert.assertEquals((Integer) 1, result.getTotalCount());
+    }
+
+    @Test
+    public void fetchClinicalDataNullAttributeSummaryProjection() throws Exception {
 
         List<String> studyIds = new ArrayList<>();
         studyIds.add("study_tcga_pub");
-        studyIds.add("acc_tcga");
-        List<String> patientIds = new ArrayList<>();
-        patientIds.add("TCGA-A1-A0SB");
-        patientIds.add("TCGA-A1-B0SO");
-        BaseMeta result = clinicalDataMyBatisRepository.getMetaPatientClinicalData(studyIds, patientIds, null);
+        studyIds.add("study_tcga_pub");
+        List<String> sampleIds = new ArrayList<>();
+        sampleIds.add("TCGA-A1-A0SB-01");
+        sampleIds.add("TCGA-A1-A0SD-01");
+        List<ClinicalData> result = clinicalDataMyBatisRepository.fetchClinicalData(studyIds, sampleIds, null,
+                PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE, "SUMMARY");
 
-        Assert.assertEquals((Integer) 7, result.getTotalCount());
+        Assert.assertEquals(6, result.size());
+        ClinicalData data = result.get(0);
+        Assert.assertEquals("OTHER_SAMPLE_ID", data.getAttrId());
+        Assert.assertEquals("5C631CE8-F96A-4C35-A459-556FC4AB21E1", data.getAttrValue());
+        Assert.assertEquals((Integer) 1, data.getInternalId());
+        Assert.assertNull(data.getClinicalAttribute());
+    }
+
+    @Test
+    public void fetchMetaClinicalDataNullAttribute() throws Exception {
+
+        List<String> studyIds = new ArrayList<>();
+        studyIds.add("study_tcga_pub");
+        studyIds.add("study_tcga_pub");
+        List<String> sampleIds = new ArrayList<>();
+        sampleIds.add("TCGA-A1-A0SB-01");
+        sampleIds.add("TCGA-A1-A0SD-01");
+        BaseMeta result = clinicalDataMyBatisRepository.fetchMetaClinicalData(studyIds, sampleIds, null,
+                PersistenceConstants.SAMPLE_CLINICAL_DATA_TYPE);
+
+        Assert.assertEquals((Integer) 6, result.getTotalCount());
     }
 }
