+++ service/src/test/java/org/cbioportal/service/impl/BaseServiceImplTest.java
@@ -12,6 +12,6 @@ public class BaseServiceImplTest {
     public static final String SAMPLE_ID = "sample_id";
     public static final String PATIENT_ID = "patient_id";
     public static final String ATTRIBUTE_ID = "attributeId";
-    public static final String PATIENT_CLINICAL_DATA_TYPE = "PATIENT";
     public static final String CANCER_TYPE_ID = "cancer_type_id";
+    public static final String CLINICAL_DATA_TYPE = "clinical_data_type";
 }
