+++ service/src/test/java/org/cbioportal/service/impl/ClinicalDataServiceImplTest.java
@@ -1,10 +1,8 @@
 package org.cbioportal.service.impl;
 
 import junit.framework.Assert;
-import org.cbioportal.model.PatientClinicalData;
-import org.cbioportal.model.SampleClinicalData;
+import org.cbioportal.model.ClinicalData;
 import org.cbioportal.model.meta.BaseMeta;
-import org.cbioportal.model.summary.ClinicalDataSummary;
 import org.cbioportal.persistence.ClinicalDataRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -14,7 +12,6 @@ import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -29,15 +26,15 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
     @Test
     public void getAllClinicalDataOfSampleInStudy() throws Exception {
 
-        List<SampleClinicalData> expectedSampleClinicalDataList = new ArrayList<>();
-        SampleClinicalData sampleClinicalData = new SampleClinicalData();
+        List<ClinicalData> expectedSampleClinicalDataList = new ArrayList<>();
+        ClinicalData sampleClinicalData = new ClinicalData();
         expectedSampleClinicalDataList.add(sampleClinicalData);
 
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfSampleInStudy(Arrays.asList(STUDY_ID),
-                Arrays.asList(SAMPLE_ID), ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
+        Mockito.when(clinicalDataRepository.getAllClinicalDataOfSampleInStudy(STUDY_ID, SAMPLE_ID, ATTRIBUTE_ID,
+                PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
                 .thenReturn(expectedSampleClinicalDataList);
 
-        List<SampleClinicalData> result = clinicalDataService.getAllClinicalDataOfSampleInStudy(STUDY_ID, SAMPLE_ID,
+        List<ClinicalData> result = clinicalDataService.getAllClinicalDataOfSampleInStudy(STUDY_ID, SAMPLE_ID,
                 ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION);
 
         Assert.assertEquals(expectedSampleClinicalDataList, result);
@@ -48,8 +45,8 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
 
         BaseMeta expectedBaseMeta = new BaseMeta();
 
-        Mockito.when(clinicalDataRepository.getMetaSampleClinicalData(Arrays.asList(STUDY_ID), Arrays.asList(SAMPLE_ID),
-                ATTRIBUTE_ID)).thenReturn(expectedBaseMeta);
+        Mockito.when(clinicalDataRepository.getMetaSampleClinicalData(STUDY_ID, SAMPLE_ID, ATTRIBUTE_ID))
+                .thenReturn(expectedBaseMeta);
 
         BaseMeta result = clinicalDataService.getMetaSampleClinicalData(STUDY_ID, SAMPLE_ID, ATTRIBUTE_ID);
 
@@ -59,15 +56,15 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
     @Test
     public void getAllClinicalDataOfPatientInStudy() throws Exception {
 
-        List<PatientClinicalData> expectedPatientClinicalDataList = new ArrayList<>();
-        PatientClinicalData patientClinicalData = new PatientClinicalData();
+        List<ClinicalData> expectedPatientClinicalDataList = new ArrayList<>();
+        ClinicalData patientClinicalData = new ClinicalData();
         expectedPatientClinicalDataList.add(patientClinicalData);
 
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfPatientInStudy(Arrays.asList(STUDY_ID),
-                Arrays.asList(PATIENT_ID), ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
+        Mockito.when(clinicalDataRepository.getAllClinicalDataOfPatientInStudy(STUDY_ID, PATIENT_ID, ATTRIBUTE_ID,
+                PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
                 .thenReturn(expectedPatientClinicalDataList);
 
-        List<PatientClinicalData> result = clinicalDataService.getAllClinicalDataOfPatientInStudy(STUDY_ID, PATIENT_ID,
+        List<ClinicalData> result = clinicalDataService.getAllClinicalDataOfPatientInStudy(STUDY_ID, PATIENT_ID,
                 ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION);
 
         Assert.assertEquals(expectedPatientClinicalDataList, result);
@@ -78,8 +75,8 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
 
         BaseMeta expectedBaseMeta = new BaseMeta();
 
-        Mockito.when(clinicalDataRepository.getMetaPatientClinicalData(Arrays.asList(STUDY_ID),
-                Arrays.asList(PATIENT_ID), ATTRIBUTE_ID)).thenReturn(expectedBaseMeta);
+        Mockito.when(clinicalDataRepository.getMetaPatientClinicalData(STUDY_ID, PATIENT_ID, ATTRIBUTE_ID))
+                .thenReturn(expectedBaseMeta);
 
         BaseMeta result = clinicalDataService.getMetaPatientClinicalData(STUDY_ID, PATIENT_ID, ATTRIBUTE_ID);
 
@@ -87,90 +84,35 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
     }
 
     @Test
-    public void getAllClinicalDataInStudySampleClinicalDataType() throws Exception {
+    public void getAllClinicalDataInStudy() throws Exception {
 
-        List<SampleClinicalData> expectedSampleClinicalDataList = new ArrayList<>();
-        SampleClinicalData sampleClinicalData = new SampleClinicalData();
+        List<ClinicalData> expectedSampleClinicalDataList = new ArrayList<>();
+        ClinicalData sampleClinicalData = new ClinicalData();
         expectedSampleClinicalDataList.add(sampleClinicalData);
 
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfSampleInStudy(Arrays.asList(STUDY_ID), null,
-                ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
+        Mockito.when(clinicalDataRepository.getAllClinicalDataInStudy(STUDY_ID, ATTRIBUTE_ID,
+                CLINICAL_DATA_TYPE, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
                 .thenReturn(expectedSampleClinicalDataList);
 
-        List<? extends ClinicalDataSummary> result = clinicalDataService.getAllClinicalDataInStudy(STUDY_ID,
-                ATTRIBUTE_ID, ClinicalDataServiceImpl.SAMPLE_CLINICAL_DATA_TYPE, PROJECTION, PAGE_SIZE, PAGE_NUMBER,
+        List<ClinicalData> result = clinicalDataService.getAllClinicalDataInStudy(STUDY_ID,
+                ATTRIBUTE_ID, CLINICAL_DATA_TYPE, PROJECTION, PAGE_SIZE, PAGE_NUMBER,
                 SORT, DIRECTION);
 
         Assert.assertEquals(expectedSampleClinicalDataList, result);
     }
 
     @Test
-    public void getAllClinicalDataInStudyPatientClinicalDataType() throws Exception {
-
-        List<PatientClinicalData> expectedPatientClinicalDataList = new ArrayList<>();
-        PatientClinicalData patientClinicalData = new PatientClinicalData();
-        expectedPatientClinicalDataList.add(patientClinicalData);
-
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfPatientInStudy(Arrays.asList(STUDY_ID), null,
-                ATTRIBUTE_ID, PROJECTION, PAGE_SIZE, PAGE_NUMBER, SORT, DIRECTION))
-                .thenReturn(expectedPatientClinicalDataList);
-
-        List<? extends ClinicalDataSummary> result = clinicalDataService.getAllClinicalDataInStudy(STUDY_ID,
-                ATTRIBUTE_ID, PATIENT_CLINICAL_DATA_TYPE, PROJECTION, PAGE_SIZE, PAGE_NUMBER,
-                SORT, DIRECTION);
-
-        Assert.assertEquals(expectedPatientClinicalDataList, result);
-    }
-
-    @Test
-    public void getMetaAllClinicalDataSampleClinicalDataType() throws Exception {
+    public void getMetaAllClinicalData() throws Exception {
 
         BaseMeta expectedBaseMeta = new BaseMeta();
         expectedBaseMeta.setTotalCount(5);
 
-        Mockito.when(clinicalDataRepository.getMetaSampleClinicalData(Arrays.asList(STUDY_ID), null, ATTRIBUTE_ID))
+        Mockito.when(clinicalDataRepository.getMetaAllClinicalData(STUDY_ID, ATTRIBUTE_ID, CLINICAL_DATA_TYPE))
                 .thenReturn(expectedBaseMeta);
 
-        BaseMeta result = clinicalDataService.getMetaAllClinicalData(STUDY_ID, ATTRIBUTE_ID,
-                ClinicalDataServiceImpl.SAMPLE_CLINICAL_DATA_TYPE);
+        BaseMeta result = clinicalDataService.getMetaAllClinicalData(STUDY_ID, ATTRIBUTE_ID, CLINICAL_DATA_TYPE);
 
         Assert.assertEquals((Integer) 5, result.getTotalCount());
-    }
-
-    @Test
-    public void getMetaAllClinicalDataPatientClinicalDataType() throws Exception {
-
-        BaseMeta expectedBaseMeta = new BaseMeta();
-        expectedBaseMeta.setTotalCount(5);
-
-        Mockito.when(clinicalDataRepository.getMetaPatientClinicalData(Arrays.asList(STUDY_ID), null, ATTRIBUTE_ID))
-                .thenReturn(expectedBaseMeta);
-
-        BaseMeta result = clinicalDataService.getMetaAllClinicalData(STUDY_ID, ATTRIBUTE_ID,
-                PATIENT_CLINICAL_DATA_TYPE);
-
-        Assert.assertEquals((Integer) 5, result.getTotalCount());
-    }
-
-    @Test
-    public void fetchClinicalDataSampleClinicalDataType() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add(STUDY_ID);
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add(SAMPLE_ID);
-
-        List<SampleClinicalData> expectedSampleClinicalDataList = new ArrayList<>();
-        SampleClinicalData sampleClinicalData = new SampleClinicalData();
-        expectedSampleClinicalDataList.add(sampleClinicalData);
-
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfSampleInStudy(studyIds, sampleIds, ATTRIBUTE_ID,
-                PROJECTION, 0, 0, null, null)).thenReturn(expectedSampleClinicalDataList);
-
-        List<? extends ClinicalDataSummary> result = clinicalDataService.fetchClinicalData(studyIds, sampleIds,
-                ATTRIBUTE_ID, ClinicalDataServiceImpl.SAMPLE_CLINICAL_DATA_TYPE, PROJECTION);
-
-        Assert.assertEquals(expectedSampleClinicalDataList, result);
     }
 
     @Test
@@ -181,37 +123,17 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
         List<String> patientIds = new ArrayList<>();
         patientIds.add(PATIENT_ID);
 
-        List<PatientClinicalData> expectedPatientClinicalDataList = new ArrayList<>();
-        PatientClinicalData patientClinicalData = new PatientClinicalData();
+        List<ClinicalData> expectedPatientClinicalDataList = new ArrayList<>();
+        ClinicalData patientClinicalData = new ClinicalData();
         expectedPatientClinicalDataList.add(patientClinicalData);
 
-        Mockito.when(clinicalDataRepository.getAllClinicalDataOfPatientInStudy(studyIds, patientIds, ATTRIBUTE_ID,
-                PROJECTION, 0, 0, null, null)).thenReturn(expectedPatientClinicalDataList);
+        Mockito.when(clinicalDataRepository.fetchClinicalData(studyIds, patientIds, ATTRIBUTE_ID, CLINICAL_DATA_TYPE,
+                PROJECTION)).thenReturn(expectedPatientClinicalDataList);
 
-        List<? extends ClinicalDataSummary> result = clinicalDataService.fetchClinicalData(studyIds, patientIds,
-                ATTRIBUTE_ID, PATIENT_CLINICAL_DATA_TYPE, PROJECTION);
+        List<ClinicalData> result = clinicalDataService.fetchClinicalData(studyIds, patientIds,
+                ATTRIBUTE_ID, CLINICAL_DATA_TYPE, PROJECTION);
 
         Assert.assertEquals(expectedPatientClinicalDataList, result);
-    }
-
-    @Test
-    public void fetchMetaClinicalDataSampleClinicalDataType() throws Exception {
-
-        List<String> studyIds = new ArrayList<>();
-        studyIds.add(STUDY_ID);
-        List<String> sampleIds = new ArrayList<>();
-        sampleIds.add(SAMPLE_ID);
-
-        BaseMeta expectedBaseMeta = new BaseMeta();
-        expectedBaseMeta.setTotalCount(5);
-
-        Mockito.when(clinicalDataRepository.getMetaSampleClinicalData(studyIds, sampleIds, ATTRIBUTE_ID))
-                .thenReturn(expectedBaseMeta);
-
-        BaseMeta result = clinicalDataService.fetchMetaClinicalData(studyIds, sampleIds, ATTRIBUTE_ID,
-                ClinicalDataServiceImpl.SAMPLE_CLINICAL_DATA_TYPE);
-
-        Assert.assertEquals((Integer) 5, result.getTotalCount());
     }
 
     @Test
@@ -225,11 +147,11 @@ public class ClinicalDataServiceImplTest extends BaseServiceImplTest {
         BaseMeta expectedBaseMeta = new BaseMeta();
         expectedBaseMeta.setTotalCount(5);
 
-        Mockito.when(clinicalDataRepository.getMetaPatientClinicalData(studyIds, patientIds, ATTRIBUTE_ID))
-                .thenReturn(expectedBaseMeta);
+        Mockito.when(clinicalDataRepository.fetchMetaClinicalData(studyIds, patientIds, ATTRIBUTE_ID,
+                CLINICAL_DATA_TYPE)).thenReturn(expectedBaseMeta);
 
         BaseMeta result = clinicalDataService.fetchMetaClinicalData(studyIds, patientIds, ATTRIBUTE_ID,
-                PATIENT_CLINICAL_DATA_TYPE);
+                CLINICAL_DATA_TYPE);
 
         Assert.assertEquals((Integer) 5, result.getTotalCount());
     }
