+++ web/src/main/java/org/cbioportal/web/GeneController.java
@@ -3,7 +3,6 @@ package org.cbioportal.web;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.cbioportal.model.Gene;
-import org.cbioportal.model.summary.GeneSummary;
 import org.cbioportal.web.parameter.PagingConstants;
 import org.cbioportal.web.parameter.Projection;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +21,7 @@ public class GeneController {
 
     @RequestMapping(value = "/genes", method = RequestMethod.GET)
     @ApiOperation("Get all genes")
-    public ResponseEntity<List<? extends GeneSummary>> getAllGenes(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<Gene>> getAllGenes(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                    @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                    @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber) {
 
@@ -31,14 +30,14 @@ public class GeneController {
 
     @RequestMapping(value = "/genes/{entrezGeneId}", method = RequestMethod.GET)
     @ApiOperation("Get a gene")
-    public ResponseEntity<GeneSummary> getGene(@PathVariable String entrezGeneId) {
+    public ResponseEntity<Gene> getGene(@PathVariable String entrezGeneId) {
 
         throw new UnsupportedOperationException();
     }
 
     @RequestMapping(value = "/genes/fetch", method = RequestMethod.POST)
     @ApiOperation("Fetch genes by ID")
-    public ResponseEntity<List<? extends GeneSummary>> fetchGenes(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<Gene>> fetchGenes(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                   @RequestBody List<String> entrezGeneIds) {
 
         throw new UnsupportedOperationException();
@@ -46,7 +45,7 @@ public class GeneController {
 
     @RequestMapping(value = "/genes/query", method = RequestMethod.POST)
     @ApiOperation("Query genes by example")
-    public ResponseEntity<List<? extends GeneSummary>> queryGenesByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<Gene>> queryGenesByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                            @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                            @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber,
                                                                            @RequestBody Gene exampleGene) {
