+++ web/src/main/java/org/cbioportal/web/GeneticDataController.java
@@ -3,7 +3,6 @@ package org.cbioportal.web;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.cbioportal.model.GeneticData;
-import org.cbioportal.model.summary.GeneticDataSummary;
 import org.cbioportal.web.parameter.PagingConstants;
 import org.cbioportal.web.parameter.Projection;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +21,7 @@ public class GeneticDataController {
 
     @RequestMapping(value = "/studies/{studyId}/samples/{sampleId}/genetic-data", method = RequestMethod.GET)
     @ApiOperation("Get all genetic data of a sample in a study")
-    public ResponseEntity<List<? extends GeneticDataSummary>> getAllGeneticDataInSampleInStudy(@PathVariable String studyId,
+    public ResponseEntity<List<GeneticData>> getAllGeneticDataInSampleInStudy(@PathVariable String studyId,
                                                                                                @PathVariable String sampleId,
                                                                                                @RequestParam String geneticProfileId,
                                                                                                @RequestParam(defaultValue = "SUMMARY") Projection projection,
@@ -34,7 +33,7 @@ public class GeneticDataController {
 
     @RequestMapping(value = "/studies/{studyId}/patients/{patientId}/genetic-data", method = RequestMethod.GET)
     @ApiOperation("Get all genetic data of a patient in a study")
-    public ResponseEntity<List<? extends GeneticDataSummary>> getAllGeneticDataInPatientInStudy(@PathVariable String studyId,
+    public ResponseEntity<List<GeneticData>> getAllGeneticDataInPatientInStudy(@PathVariable String studyId,
                                                                                                 @PathVariable String patientId,
                                                                                                 @RequestParam String geneticProfileId,
                                                                                                 @RequestParam(defaultValue = "SUMMARY") Projection projection,
@@ -46,7 +45,7 @@ public class GeneticDataController {
 
     @RequestMapping(value = "/genetic-profiles/{geneticProfileId}/genetic-data", method = RequestMethod.GET)
     @ApiOperation("Get all genetic data in a genetic profile")
-    public ResponseEntity<List<? extends GeneticDataSummary>> getAllGeneticDataInGeneticProfile(@PathVariable String geneticProfileId,
+    public ResponseEntity<List<GeneticData>> getAllGeneticDataInGeneticProfile(@PathVariable String geneticProfileId,
                                                                                                 @RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                                                 @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                                                 @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber) {
@@ -56,7 +55,7 @@ public class GeneticDataController {
 
     @RequestMapping(value = "/genetic-data/query", method = RequestMethod.POST)
     @ApiOperation("Query genetic data by example")
-    public ResponseEntity<List<? extends GeneticDataSummary>> queryGeneticDataByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<GeneticData>> queryGeneticDataByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                                         @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                                         @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber,
                                                                                         @RequestBody GeneticData exampleGenericData) {
