+++ web/src/main/java/org/cbioportal/web/PatientController.java
@@ -3,7 +3,6 @@ package org.cbioportal.web;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.cbioportal.model.Patient;
-import org.cbioportal.model.summary.PatientSummary;
 import org.cbioportal.web.parameter.PagingConstants;
 import org.cbioportal.web.parameter.PatientIdentifier;
 import org.cbioportal.web.parameter.Projection;
@@ -23,7 +22,7 @@ public class PatientController {
 
     @RequestMapping(value = "/studies/{studyId}/patients", method = RequestMethod.GET)
     @ApiOperation("Get all patients in a study")
-    public ResponseEntity<List<? extends PatientSummary>> getAllPatientsInStudy(@PathVariable String studyId,
+    public ResponseEntity<List<Patient>> getAllPatientsInStudy(@PathVariable String studyId,
                                                                                 @RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                                 @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                                 @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber) {
@@ -33,7 +32,7 @@ public class PatientController {
 
     @RequestMapping(value = "/studies/{studyId}/patients/{patientId}", method = RequestMethod.GET)
     @ApiOperation("Get a patient in a study")
-    public ResponseEntity<PatientSummary> getPatientInStudy(@PathVariable String studyId,
+    public ResponseEntity<Patient> getPatientInStudy(@PathVariable String studyId,
                                                             @PathVariable String patientId) {
 
         throw new UnsupportedOperationException();
@@ -41,7 +40,7 @@ public class PatientController {
 
     @RequestMapping(value = "/patients/fetch", method = RequestMethod.POST)
     @ApiOperation("Fetch patients by ID")
-    public ResponseEntity<List<? extends PatientSummary>> fetchPatients(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<Patient>> fetchPatients(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                         @RequestBody List<PatientIdentifier> patientIdentifiers) {
 
         throw new UnsupportedOperationException();
@@ -49,12 +48,11 @@ public class PatientController {
 
     @RequestMapping(value = "/patients/query", method = RequestMethod.POST)
     @ApiOperation("Query patients by example")
-    public ResponseEntity<List<? extends PatientSummary>> queryPatientsByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
+    public ResponseEntity<List<Patient>> queryPatientsByExample(@RequestParam(defaultValue = "SUMMARY") Projection projection,
                                                                                  @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_SIZE) Integer pageSize,
                                                                                  @RequestParam(defaultValue = PagingConstants.DEFAULT_PAGE_NUMBER) Integer pageNumber,
                                                                                  @RequestBody Patient examplePatient) {
 
         throw new UnsupportedOperationException();
     }
-
 }
