+++ web/src/main/java/org/cbioportal/web/mixin/ClinicalAttributeMixin.java
@@ -1,9 +1,18 @@
 package org.cbioportal.web.mixin;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.cbioportal.model.CancerStudy;
-import org.cbioportal.web.mixin.summary.ClinicalAttributeSummaryMixin;
 
-public class ClinicalAttributeMixin extends ClinicalAttributeSummaryMixin {
+public class ClinicalAttributeMixin {
 
+    private String attrId;
+    private String displayName;
+    private String description;
+    private String datatype;
+    private Boolean patientAttribute;
+    private String priority;
+    @JsonIgnore
+    private Integer cancerStudyId;
+    @JsonIgnore
     private CancerStudy cancerStudy;
 }
