+++ web/src/main/java/org/cbioportal/web/mixin/ClinicalDataMixin.java
@@ -1,10 +1,10 @@
-package org.cbioportal.web.mixin.summary;
+package org.cbioportal.web.mixin;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.cbioportal.model.ClinicalAttribute;
 
-public class ClinicalDataSummaryMixin {
+public class ClinicalDataMixin {
 
     @JsonIgnore
     private Integer internalId;
@@ -12,4 +12,5 @@ public class ClinicalDataSummaryMixin {
     private String stableId;
     private String attrId;
     private String attrValue;
+    private ClinicalAttribute clinicalAttribute;
 }
