+++ web/src/main/java/org/cbioportal/web/mixin/GeneMixin.java
@@ -1,10 +1,14 @@
 package org.cbioportal.web.mixin;
 
-import org.cbioportal.web.mixin.summary.GeneSummaryMixin;
-
 import java.util.List;
 
-public class GeneMixin extends GeneSummaryMixin {
+public class GeneMixin {
 
+    private Integer entrezGeneId;
+    private String hugoGeneSymbol;
+    private String type;
+    private String cytoband;
+    private Integer length;
     private List<String> aliases;
+    private String chromosome;
 }
