+++ web/src/main/java/org/cbioportal/web/mixin/GeneticDataMixin.java
@@ -1,12 +1,23 @@
 package org.cbioportal.web.mixin;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.cbioportal.model.Gene;
 import org.cbioportal.model.GeneticProfile;
 import org.cbioportal.model.Sample;
-import org.cbioportal.web.mixin.summary.GeneticDataSummaryMixin;
 
-public class GeneticDataMixin extends GeneticDataSummaryMixin {
+public class GeneticDataMixin {
 
+    @JsonIgnore
+    private Integer geneticProfileId;
+    @JsonProperty("geneticProfileId")
+    private String geneticProfileStableId;
+    private Integer entrezGeneId;
+    @JsonIgnore
+    private Integer sampleId;
+    @JsonProperty("sampleStableId")
+    private String sampleStableId;
+    private String value;
     private GeneticProfile geneticProfile;
     private Gene gene;
     private Sample sample;
