+++ web/src/main/java/org/cbioportal/web/mixin/GeneticProfileMixin.java
@@ -1,9 +1,22 @@
 package org.cbioportal.web.mixin;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.cbioportal.model.CancerStudy;
-import org.cbioportal.web.mixin.summary.GeneticProfileSummaryMixin;
 
-public class GeneticProfileMixin extends GeneticProfileSummaryMixin{
+public class GeneticProfileMixin {
 
+    @JsonIgnore
+    private Integer geneticProfileId;
+    private String stableId;
+    @JsonIgnore
+    private Integer cancerStudyId;
+    @JsonProperty("cancerStudyId")
+    private String cancerStudyIdentifier;
+    private String geneticAlterationType;
+    private String datatype;
+    private String name;
+    private String description;
+    private Boolean showProfileInAnalysisTab;
     private CancerStudy cancerStudy;
 }
