+++ web/src/main/java/org/cbioportal/web/mixin/PatientMixin.java
@@ -1,10 +1,17 @@
 package org.cbioportal.web.mixin;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.cbioportal.model.CancerStudy;
-import org.cbioportal.web.mixin.summary.PatientSummaryMixin;
 
+public class PatientMixin {
 
-public class PatientMixin extends PatientSummaryMixin {
-
+    @JsonIgnore
+    private Integer internalId;
+    private String stableId;
+    @JsonIgnore
+    private Integer cancerStudyId;
+    @JsonProperty("cancerStudyId")
+    private String cancerStudyIdentifier;
     private CancerStudy cancerStudy;
 }
