--- web/src/main/java/org/cbioportal/web/mixin/summary/GeneSummaryMixin.java
@@ -1,10 +1,0 @@
-package org.cbioportal.web.mixin.summary;
-
-public class GeneSummaryMixin {
-
-    private Integer entrezGeneId;
-    private String hugoGeneSymbol;
-    private String type;
-    private String cytoband;
-    private Integer length;
-}
