--- web/src/main/java/org/cbioportal/web/mixin/summary/GeneticProfileSummaryMixin.java
@@ -1,20 +1,0 @@
-package org.cbioportal.web.mixin.summary;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class GeneticProfileSummaryMixin {
-
-    @JsonIgnore
-    private Integer geneticProfileId;
-    private String stableId;
-    @JsonIgnore
-    private Integer cancerStudyId;
-    @JsonProperty("cancerStudyId")
-    private String cancerStudyIdentifier;
-    private String geneticAlterationType;
-    private String datatype;
-    private String name;
-    private String description;
-    private Boolean showProfileInAnalysisTab;
-}
